kafka-structured/services/scaling-controller/requirements.txt
kafka-python==2.0.2
kubernetes==28.1.0
